+++ src/main/java/ar/edu/utn/frbb/tup/persistence/entity/CuentaEntity.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 public class CuentaEntity extends BaseEntity{
     String nombre;
     LocalDateTime fechaCreacion;
-    int balance;
+    double balance;
     String tipoCuenta;
     Long titular;
     long numeroCuenta;
@@ -48,11 +48,11 @@ public class CuentaEntity extends BaseEntity{
         this.fechaCreacion = fechaCreacion;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
